Introduced MainViewModel to handle UI logic outside of activity
The MainActivity is responsible for
- displaying quotes to the user
- reacting to refresh button clicks

MainViewModel now handles the UI logic which includes:
- notifying the activity when loading a quote
- notifiying the activity when the quote is ready to be displayed.
- Using a separate thread to prevent blocking the UI thread

Because MainViewModel out lives the state of MainActivity, our activity can now handle configuration changes (rotating the screen) for free.
